requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Authentication
Flask-Login==0.6.3

# Email
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Google OAuth
Authlib==1.3.0

# AWS S3
boto3==1.42.42

# API REST
flask-restx==1.3.2
PyJWT==2.11.0
